Phase 2.1: split db/schema.ts into per-domain modules
897-line schema.ts becomes src/db/schema/ with tables grouped by domain
(courses, topics, dailies, routines, tasks, tags, radar, users, enums)
and all relations() in a single relations.ts to keep the module graph
acyclic. Declarations moved verbatim; the index barrel preserves all 71
existing @/db/schema imports unchanged.

Verified: drizzle-kit push produces identical statements for the old and
new schema; runtime import + endpoint smoke against a live DB pass.
