+++ packages/middleware/drizzle.config.ts
@@ -4,7 +4,7 @@ import { defineConfig } from "drizzle-kit";
 
 export default defineConfig({
   out: "./drizzle",
-  schema: "./src/db/schema.ts",
+  schema: "./src/db/schema/index.ts",
   dialect: "postgresql",
   dbCredentials: {
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion,no-undef
