+++ packages/middleware/src/db/schema/index.ts
@@ -1,0 +1,13 @@
+// Drizzle schema, split by domain. Table modules may import each other's
+// tables for FK references (lazy callbacks, no cycles); all relations()
+// declarations live in relations.ts.
+export * from "./courses";
+export * from "./dailies";
+export * from "./enums";
+export * from "./radar";
+export * from "./relations";
+export * from "./routines";
+export * from "./tags";
+export * from "./tasks";
+export * from "./topics";
+export * from "./users";
