+++ packages/middleware/src/db/schema/users.ts
@@ -1,0 +1,13 @@
+import { boolean, integer, pgTable, varchar } from "drizzle-orm/pg-core";
+
+export const usersTable = pgTable("users", {
+  id: varchar().primaryKey(),
+  name: varchar({
+    length: 255,
+  }).notNull(),
+  age: integer().notNull(),
+  test: boolean(),
+  email: varchar({
+    length: 255,
+  }).notNull().unique(),
+});
